Add the first side-effectful filters behind --enable-filters.
`--enable-filters` takes a comma-separated list of non-standard filters to
enable in `p2`. Notably this implements `write_file` and `make_dirs`, which
allow a template execution to have filesystem side-effects.

The write_file filter allows using a single template to output multiple files,
including customization of filenames from other templated values.

The write_file filter returns it's literal input, so stdout / file output can
be used as a log of all templated values.

make_dirs will create the directory path given as it's argument. It is the
natural companion of write_files, but separate in order to be explicit about
expected template operations.

`--enable-noop-filters` loads all custom filters but disables their
side-effects, in order to allow easy debugging since filter use is not
necessarily obvious.

This feature is considered experimental at the moment, and may be deprecated
or changed in future releases.